src/config.py
import os
from pathlib import Path
from dotenv import load_dotenv

# Load .env file
load_dotenv()

class Settings:
    SPOTIPY_CLIENT_ID = os.getenv("SPOTIPY_CLIENT_ID")
    SPOTIPY_CLIENT_SECRET = os.getenv("SPOTIPY_CLIENT_SECRET")
    SPOTIPY_REDIRECT_URI = os.getenv("SPOTIPY_REDIRECT_URI", "http://localhost:8888/callback")
    
    @property
    def is_spotify_configured(self) -> bool:
        return all([self.SPOTIPY_CLIENT_ID, self.SPOTIPY_CLIENT_SECRET])

settings = Settings()
