requirements.txt
spotipy
typer
python-dotenv
rich
